api/requirements.txt
Flask==3.0.3
redis==7.0.1
pydantic-to-typescript==2.0.0
rq==2.6.0
boto3==1.40.71